requirements.txt
# Core ML Libraries
tensorflow==2.18.0
keras==2.18.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.3.1

# Image Processing
Pillow==10.0.0
opencv-python==4.8.0.76

# Visualization
matplotlib==3.8.0
seaborn==0.12.2
plotly==5.15.0

# Data Processing
tqdm==4.65.0
pathlib2==2.3.7

# Web Interface
streamlit==1.28.0
streamlit-option-menu==0.3.6

# Jupyter
jupyter==1.0.0
ipykernel==6.25.0
ipywidgets==8.0.7

# Utilities
requests==2.31.0
urllib3==2.0.4
